Revert "Config server refactoring, part 7"
