+++ configserver/src/main/java/com/yahoo/vespa/config/server/ApplicationRepository.java
@@ -49,9 +49,10 @@ import com.yahoo.vespa.config.server.http.v2.PrepareResult;
 import com.yahoo.vespa.config.server.metrics.ApplicationMetricsRetriever;
 import com.yahoo.vespa.config.server.provision.HostProvisionerProvider;
 import com.yahoo.vespa.config.server.session.LocalSession;
-import com.yahoo.vespa.config.server.session.SessionRepository;
+import com.yahoo.vespa.config.server.session.LocalSessionRepo;
 import com.yahoo.vespa.config.server.session.PrepareParams;
 import com.yahoo.vespa.config.server.session.RemoteSession;
+import com.yahoo.vespa.config.server.session.RemoteSessionRepo;
 import com.yahoo.vespa.config.server.session.Session;
 import com.yahoo.vespa.config.server.session.SessionFactory;
 import com.yahoo.vespa.config.server.session.SilentDeployLogger;
@@ -304,7 +305,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
         if (activeSession == null) return Optional.empty();
         TimeoutBudget timeoutBudget = new TimeoutBudget(clock, timeout);
         LocalSession newSession = tenant.getSessionFactory().createSessionFromExisting(activeSession, logger, true, timeoutBudget);
-        tenant.getSessionRepository().addSession(newSession);
+        tenant.getLocalSessionRepo().addSession(newSession);
 
         return Optional.of(Deployment.unprepared(newSession, this, hostProvisioner, tenant, timeout, clock,
                                                  false /* don't validate as this is already deployed */, bootstrap));
@@ -488,7 +489,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
             Tenant tenant = tenantRepository.getTenant(applicationId.tenant());
             if (tenant == null) throw new NotFoundException("Tenant '" + applicationId.tenant() + "' not found");
             long sessionId = getSessionIdForApplication(tenant, applicationId);
-            RemoteSession session = tenant.getSessionRepo().getRemoteSession(sessionId);
+            RemoteSession session = tenant.getRemoteSessionRepo().getSession(sessionId);
             if (session == null) throw new NotFoundException("Remote session " + sessionId + " not found");
             return session.ensureApplicationLoaded().getForVersionOrLatest(version, clock.instant());
         } catch (NotFoundException e) {
@@ -525,10 +526,10 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
     }
 
     private boolean localSessionHasBeenDeleted(ApplicationId applicationId, long sessionId, Duration waitTime) {
-        SessionRepository sessionRepository = tenantRepository.getTenant(applicationId.tenant()).getSessionRepo();
+        RemoteSessionRepo remoteSessionRepo = tenantRepository.getTenant(applicationId.tenant()).getRemoteSessionRepo();
         Instant end = Instant.now().plus(waitTime);
         do {
-            if (sessionRepository.getRemoteSession(sessionId) == null) return true;
+            if (remoteSessionRepo.getSession(sessionId) == null) return true;
             try { Thread.sleep(10); } catch (InterruptedException e) { /* ignored */}
         } while (Instant.now().isBefore(end));
 
@@ -634,11 +635,11 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
                                           boolean internalRedeploy,
                                           TimeoutBudget timeoutBudget) {
         Tenant tenant = tenantRepository.getTenant(applicationId.tenant());
-        SessionRepository sessionRepository = tenant.getSessionRepository();
+        LocalSessionRepo localSessionRepo = tenant.getLocalSessionRepo();
         SessionFactory sessionFactory = tenant.getSessionFactory();
         RemoteSession fromSession = getExistingSession(tenant, applicationId);
         LocalSession session = sessionFactory.createSessionFromExisting(fromSession, logger, internalRedeploy, timeoutBudget);
-        sessionRepository.addSession(session);
+        localSessionRepo.addSession(session);
         return session.getSessionId();
     }
 
@@ -659,13 +660,13 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
         Optional<Long> activeSessionId = tenant.getApplicationRepo().activeSessionOf(applicationId);
         LocalSession session = tenant.getSessionFactory().createSession(applicationDirectory, applicationId,
                                                                         timeoutBudget, activeSessionId);
-        tenant.getSessionRepository().addSession(session);
+        tenant.getLocalSessionRepo().addSession(session);
         return session.getSessionId();
     }
 
     public void deleteExpiredLocalSessions() {
         Map<Tenant, List<LocalSession>> sessionsPerTenant = new HashMap<>();
-        tenantRepository.getAllTenants().forEach(tenant -> sessionsPerTenant.put(tenant, tenant.getSessionRepository().getSessions()));
+        tenantRepository.getAllTenants().forEach(tenant -> sessionsPerTenant.put(tenant, tenant.getLocalSessionRepo().getSessions()));
 
         Set<ApplicationId> applicationIds = new HashSet<>();
         sessionsPerTenant.values().forEach(sessionList -> sessionList.forEach(s -> applicationIds.add(s.getApplicationId())));
@@ -676,7 +677,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
             if (activeSession != null)
                 activeSessions.put(applicationId, activeSession.getSessionId());
         });
-        sessionsPerTenant.keySet().forEach(tenant -> tenant.getSessionRepository().deleteExpiredSessions(activeSessions));
+        sessionsPerTenant.keySet().forEach(tenant -> tenant.getLocalSessionRepo().deleteExpiredSessions(activeSessions));
     }
 
     public int deleteExpiredRemoteSessions(Duration expiryTime) {
@@ -686,7 +687,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
     public int deleteExpiredRemoteSessions(Clock clock, Duration expiryTime) {
         return tenantRepository.getAllTenants()
                 .stream()
-                .map(tenant -> tenant.getSessionRepo().deleteExpiredRemoteSessions(clock, expiryTime))
+                .map(tenant -> tenant.getRemoteSessionRepo().deleteExpiredSessions(clock, expiryTime))
                 .mapToInt(i -> i)
                 .sum();
     }
@@ -746,14 +747,14 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
     }
 
     private LocalSession getLocalSession(Tenant tenant, long sessionId) {
-        LocalSession session = tenant.getSessionRepository().getSession(sessionId);
+        LocalSession session = tenant.getLocalSessionRepo().getSession(sessionId);
         if (session == null) throw new NotFoundException("Session " + sessionId + " was not found");
 
         return session;
     }
 
     private RemoteSession getRemoteSession(Tenant tenant, long sessionId) {
-        RemoteSession session = tenant.getSessionRepo().getRemoteSession(sessionId);
+        RemoteSession session = tenant.getRemoteSessionRepo().getSession(sessionId);
         if (session == null) throw new NotFoundException("Session " + sessionId + " was not found");
 
         return session;
@@ -804,7 +805,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
     private RemoteSession getActiveSession(Tenant tenant, ApplicationId applicationId) {
         TenantApplications applicationRepo = tenant.getApplicationRepo();
         if (applicationRepo.activeApplications().contains(applicationId)) {
-            return tenant.getSessionRepo().getRemoteSession(applicationRepo.requireActiveSessionOf(applicationId));
+            return tenant.getRemoteSessionRepo().getSession(applicationRepo.requireActiveSessionOf(applicationId));
         }
         return null;
     }
@@ -812,7 +813,7 @@ public class ApplicationRepository implements com.yahoo.config.provision.Deploye
     public LocalSession getActiveLocalSession(Tenant tenant, ApplicationId applicationId) {
         TenantApplications applicationRepo = tenant.getApplicationRepo();
         if (applicationRepo.activeApplications().contains(applicationId)) {
-            return tenant.getSessionRepository().getSession(applicationRepo.requireActiveSessionOf(applicationId));
+            return tenant.getLocalSessionRepo().getSession(applicationRepo.requireActiveSessionOf(applicationId));
         }
         return null;
     }
