+++ configserver/src/main/java/com/yahoo/vespa/config/server/tenant/Tenant.java
@@ -6,8 +6,9 @@ import com.yahoo.path.Path;
 import com.yahoo.vespa.config.server.ReloadHandler;
 import com.yahoo.vespa.config.server.RequestHandler;
 import com.yahoo.vespa.config.server.application.TenantApplications;
+import com.yahoo.vespa.config.server.session.LocalSessionRepo;
+import com.yahoo.vespa.config.server.session.RemoteSessionRepo;
 import com.yahoo.vespa.config.server.session.SessionFactory;
-import com.yahoo.vespa.config.server.session.SessionRepository;
 import com.yahoo.vespa.curator.Curator;
 import org.apache.zookeeper.data.Stat;
 
@@ -27,9 +28,10 @@ public class Tenant implements TenantHandlerProvider {
     static final String APPLICATIONS = "applications";
 
     private final TenantName name;
+    private final RemoteSessionRepo remoteSessionRepo;
     private final Path path;
     private final SessionFactory sessionFactory;
-    private final SessionRepository sessionRepository;
+    private final LocalSessionRepo localSessionRepo;
     private final TenantApplications applicationRepo;
     private final RequestHandler requestHandler;
     private final ReloadHandler reloadHandler;
@@ -37,7 +39,8 @@ public class Tenant implements TenantHandlerProvider {
 
     Tenant(TenantName name,
            SessionFactory sessionFactory,
-           SessionRepository sessionRepository,
+           LocalSessionRepo localSessionRepo,
+           RemoteSessionRepo remoteSessionRepo,
            RequestHandler requestHandler,
            ReloadHandler reloadHandler,
            TenantApplications applicationRepo,
@@ -46,8 +49,9 @@ public class Tenant implements TenantHandlerProvider {
         this.path = TenantRepository.getTenantPath(name);
         this.requestHandler = requestHandler;
         this.reloadHandler = reloadHandler;
+        this.remoteSessionRepo = remoteSessionRepo;
         this.sessionFactory = sessionFactory;
-        this.sessionRepository = sessionRepository;
+        this.localSessionRepo = localSessionRepo;
         this.applicationRepo = applicationRepo;
         this.curator = curator;
     }
@@ -70,8 +74,13 @@ public class Tenant implements TenantHandlerProvider {
         return requestHandler;
     }
 
-    public SessionRepository getSessionRepo() {
-        return sessionRepository;
+    /**
+     * The RemoteSessionRepo for this
+     *
+     * @return repo
+     */
+    public RemoteSessionRepo getRemoteSessionRepo() {
+        return remoteSessionRepo;
     }
 
     public TenantName getName() {
@@ -86,8 +95,8 @@ public class Tenant implements TenantHandlerProvider {
         return sessionFactory;
     }
 
-    public SessionRepository getSessionRepository() {
-        return sessionRepository;
+    public LocalSessionRepo getLocalSessionRepo() {
+        return localSessionRepo;
     }
 
     @Override
@@ -131,9 +140,9 @@ public class Tenant implements TenantHandlerProvider {
      * Called by watchers as a reaction to {@link #delete()}.
      */
     void close() {
-        sessionRepository.close();              // Closes watchers and clears memory.
+        remoteSessionRepo.close();              // Closes watchers and clears memory.
         applicationRepo.close();                // Closes watchers.
-        sessionRepository.close();              // Closes watchers, clears memory, and deletes local files and ZK session state.
+        localSessionRepo.close();               // Closes watchers, clears memory, and deletes local files and ZK session state.
     }
 
     /** Deletes the tenant tree from ZooKeeper (application and session status for the tenant) and triggers {@link #close()}. */
