+++ configserver/src/main/java/com/yahoo/vespa/config/server/tenant/TenantRepository.java
@@ -14,7 +14,8 @@ import com.yahoo.vespa.config.server.RequestHandler;
 import com.yahoo.vespa.config.server.application.TenantApplications;
 import com.yahoo.vespa.config.server.deploy.TenantFileSystemDirs;
 import com.yahoo.vespa.config.server.monitoring.MetricUpdater;
-import com.yahoo.vespa.config.server.session.SessionRepository;
+import com.yahoo.vespa.config.server.session.LocalSessionRepo;
+import com.yahoo.vespa.config.server.session.RemoteSessionRepo;
 import com.yahoo.vespa.config.server.session.SessionFactory;
 import com.yahoo.vespa.curator.Curator;
 import org.apache.curator.framework.CuratorFramework;
@@ -48,7 +49,7 @@ import java.util.stream.Collectors;
  * This component will monitor the set of tenants in the config server by watching in ZooKeeper.
  * It will set up Tenant objects accordingly, which will manage the config sessions per tenant.
  * This class will read the preexisting set of tenants from ZooKeeper at startup. (For now it will also
- * create a default tenant since that will be used for APIs that do no know about tenants or have not yet
+ * create a default tenant since that will be used for API that do no know about tenants or have not yet
  * implemented support for it).
  *
  * This instance is called from two different threads, the http handler threads and the zookeeper watcher threads.
@@ -223,10 +224,15 @@ public class TenantRepository {
         if (reloadHandler == null)
             reloadHandler = applicationRepo;
         SessionFactory sessionFactory = new SessionFactory(componentRegistry, applicationRepo, applicationRepo, tenantName);
-        SessionRepository sessionRepository = new SessionRepository(tenantName, componentRegistry, sessionFactory,
-                                                                    applicationRepo, reloadHandler);
+        LocalSessionRepo localSessionRepo = new LocalSessionRepo(tenantName, componentRegistry, sessionFactory);
+        RemoteSessionRepo remoteSessionRepo = new RemoteSessionRepo(componentRegistry,
+                                                                    sessionFactory,
+                                                                    reloadHandler,
+                                                                    tenantName,
+                                                                    applicationRepo,
+                                                                    componentRegistry.getFlagSource());
         log.log(Level.INFO, "Creating tenant '" + tenantName + "'");
-        Tenant tenant = new Tenant(tenantName, sessionFactory, sessionRepository, requestHandler,
+        Tenant tenant = new Tenant(tenantName, sessionFactory, localSessionRepo, remoteSessionRepo, requestHandler,
                                    reloadHandler, applicationRepo, componentRegistry.getCurator());
         notifyNewTenant(tenant);
         tenants.putIfAbsent(tenantName, tenant);
